+++ rpms/SPECS/4.0.0/wazuh-manager-4.0.0.spec
@@ -175,12 +175,15 @@ elif ! id -g ossec > /dev/null 2>&1; then
   groupadd -r ossec
 fi
 
+# Stop the services to upgrade the package
 if [ $1 = 2 ]; then
-  if command -v systemctl > /dev/null 2>&1 && systemctl > /dev/null 2>&1; then
+  if command -v systemctl > /dev/null 2>&1 && systemctl > /dev/null 2>&1 && systemctl is-active --quiet wazuh-manager > /dev/null 2>&1; then
+    touch %{_localstatedir}/tmp/wazuh.restart
     systemctl stop wazuh-manager.service > /dev/null 2>&1
   fi
 
-  if command -v service > /dev/null 2>&1; then
+  if command -v service > /dev/null 2>&1 && service wazuh-manager status | grep "is running" > /dev/null 2>&1; then
+    touch %{_localstatedir}/tmp/wazuh.restart
     service wazuh-manager stop > /dev/null 2>&1
   fi
 fi
@@ -388,19 +391,19 @@ rm -f %{_localstatedir}/etc/shared/default/*.rpmnew
 
 %preun
 
-# Stop the services before upgrading/uninstalling the package
-# Check for systemd
-if command -v systemctl > /dev/null 2>&1 && systemctl > /dev/null 2>&1; then
-  systemctl stop wazuh-manager > /dev/null 2>&1
-# Check for SysV
-elif command -v service > /dev/null 2>&1; then
-  service wazuh-manager stop > /dev/null 2>&1
-# Anything else
-else
-  %{_localstatedir}/bin/ossec-control stop > /dev/null 2>&1
-fi
-
 if [ $1 = 0 ]; then
+
+  # Stop the services before uninstall the package
+  # Check for systemd
+  if command -v systemctl > /dev/null 2>&1 && systemctl > /dev/null 2>&1; then
+    systemctl stop wazuh-manager > /dev/null 2>&1
+  # Check for SysV
+  elif command -v service > /dev/null 2>&1; then
+    service wazuh-manager stop > /dev/null 2>&1
+  # Anything else
+  else
+    %{_localstatedir}/bin/ossec-control stop > /dev/null 2>&1
+  fi
 
   # Check for systemd
   if command -v systemctl > /dev/null 2>&1 && systemctl > /dev/null 2>&1; then
@@ -472,7 +475,20 @@ if [ $1 = 0 ];then
   rm -rf %{_localstatedir}/logs/
   rm -rf %{_localstatedir}/ruleset/
   rm -rf %{_localstatedir}/tmp
+fi
 
+# posttrans code is the last thing executed in a install/upgrade
+%posttrans
+if [ -f %{_localstatedir}/tmp/wazuh.restart ]; then
+  if command -v systemctl > /dev/null 2>&1 && systemctl > /dev/null 2>&1 ; then
+    rm -f %{_localstatedir}/tmp/wazuh.restart
+    systemctl restart wazuh-manager.service > /dev/null 2>&1
+  fi
+
+  if command -v service > /dev/null 2>&1; then
+    rm -f %{_localstatedir}/tmp/wazuh.restart
+    service wazuh-manager restart > /dev/null 2>&1
+  fi
 fi
 
 %triggerin -- glibc
